starlark: fix crash due to parameter type mismatch
This change fixes a crash during a call to a Starlark built-in
function where the argument was considered valid by the annotations
yet was not assignable to the parameter variable.

The real problem is that the annotations are too complex, leading
to inconsistency. In follow-up changes, I will eliminate both Param.type,
which is (or rather, should be) redundant w.r.t. an allowedTypes array
of length 1, and Param.noneable, which is (again, should be)
redundant w.r.t. the existence of NoneType among allowedTypes.

Details:
- Void.class as sentinel value for Param.type, meaning "the class
  of the parameter variable". The former sentinel, Object.class,
  was ambiguous.
- Change dynamic semantics to match.
- Eliminate explicit type=Object.class annotations in Bazel where
  allowedTypes is also specified. The static checker now detects this
  error because of the new sentinel.
- Make ParamType.type mandatory.
- Add regression test.
- Update Copybara's Markdown generator to correctly compute the
  effective parameter type from .allowedTypes, .type, param var type,
  noneable.

This is a breaking API change for Copybara. Sorry.

PiperOrigin-RevId: 335853473
Change-Id: Id8b86dda956457f281daa64203d608aa1e0c92d9