doc generator: save nested classes with '$'
The documentation generator will load the class later using
classLoader.loadClass(name). When name is a nested class, the
format need so be like "foo.pkg.SomeClass$Nested" instead of
"foo.pkg.SomeClass.Nested".

This change fixes the generator to generate in that format.

One thing to note is that this only works for one single
nested class. Not multi-nested classes (foo.SomeClass$Bar$Baz).
But we don't care about those cases.

Change-Id: I2d3f29eb317f7b380b114d6b09e857e19e1403be