Scaffolding for Starlark git.mirror migrations
This change adds the basic wiring to allow custom git.mirror logic.
The reason is that we want to support configurable rebases/merges.

The general idea is that the actions will be in charge of
fetching/pushing what they need (but won't be allowed to fetch/push
anything not in the refspec and from the defined origin/destination).

This first change only add the wiring, and still doesn't support
functions with parameters.

Side change: Moved the non-fast-forward check to GitRepository to
abstract the callers from making string checks. Exception probably
should be VE, but we don't have retriable VE, so better RepoException
for now.

BUG=185208243
Change-Id: I31abd0075073d68fa5d109943437559359379484