Use GitHub API to detect presence of merge commit
It is possible for a stale merge commit (against a previous version of the baseline) to exist at "refs/pull/<PR #>/merge". Therefore, when determining whether an up-to-date merge commit actually exists, it is better to use the explicit "mergeable" bit from the GitHub API rather than using the existence of such a ref.

PiperOrigin-RevId: 371388134
Change-Id: Iea279dbef02606e69d6ec65666764ba631883b0f