starlark git.mirror: Basic merge code
After this change we are able to fetch references from the destination
and perform a merge between the origin & destination.

Most of the operations return a value on error, so that Starlark code
can recover in case of error (For example control what to do when
destination fetch doesn't exist or merge fails).

Most of the functions work with the bare repo to avoid failures trying
to fetch on the same branch for non-bare repos errors [1] and not making
the user to use different branches to avoid it. The added benefit is
that if we add more operations that need the worktree, those operations
are more isolated.

[1] Git error: Refusing to fetch into current branch refs/heads/master of non-bare repository

BUG=185208243

Change-Id: I9e00ed5428b1aca9f8cbfd2cdf15a7bc5c1c9c60