Simplify TreeState
We had 2 TreeState implementation classes that were doing the same
thing. By simplifying to a single class, we can also mutate the
instance in-place, preventing us from having to copy the TreeState
object (as well as the wrapping TransformWork instance). This change
will make it easier for us to make use of the cache more often.

PiperOrigin-RevId: 376174288
Change-Id: I3f932fd3e8595da5a961be8483dc44bcae88750e