Modify how Transformations report no-ops
Previously, a no-op transformation would directly throw a
`VoidOperationException`. Now, it returns a TransformationStatus object
representing either a 'success' or a 'no-op' so that the Sequence in
which it is contained can handle a potential no-op appropriately. Even
Sequence transformations themselves don't directly handle no-ops; they
also return a TransformationStatus object. In this way, no-ops bubble
up the transformation tree and are handled at the root.

The exception to this behavior is transforms which exist inside
`core.transform(ignore_noop=False, ...)`, which (as the always have)
immediately throw `VoidOperationException`s, rather than bubbling up to
the root of the transformation tree.

The `--ignore-noop` command-line flag continues to skip 100% of no-op
related exceptions, even those mentioned in the previous paragraph.

PiperOrigin-RevId: 377088730
Change-Id: I679a4df843401bf4bc4d8147cfba5cc16c5d06a3