Handle GitHub null mergeable edge case
We use the value of 'mergeable' from the GitHub PR API response to tell
us whether a merge ref exists. This field can take on 3 possible values:
true, to indicate that the ref definitely does exist, false when it does
not exist, and null which essentially just means maybe. (See
https://docs.github.com/en/rest/reference/pulls#get-a-pull-request.)

When 'mergeable' has a value of null, we try to fetch the merge ref in
case it does exist. When run with '--force', we will fall back to a
second fetch in the event that the merge ref was absent.

PiperOrigin-RevId: 380892588
Change-Id: I9206b29fc2bd5dac274edc372b580f1fb236eef8