Distinguish between an empty change and a redundant change
Create a new Type of DestinationEffect to correspond to the case that
a duplicate change is re-imported when `allow_empty_diff[_patchset]` is
set to False. This case was formerly classified as a NOOP, but in truth
a change still exists relative to the destination (it's just that this
change has already been imported into a pending change in the
destination), so we instead use `NOOP_AGAINST_PENDING_CHANGE`.

BUG=192013669
PiperOrigin-RevId: 384748081
Change-Id: Idefaf761f34aaad803f044c417b409328a6442f6