GitHub PR#190 Add support for options specified in `//:.bazelrc.user`
--
9668011cc87666278573550b0a2c537f3bcdf127 by Benjamin Denhartog <ben@sudoforge.com>:

Add support for options specified in `//:.bazelrc.user`

With this change, users can specify options for Bazel in a configuration
file at `//:.bazelrc.user`. The options specified in this file will
be merged with options specified in `//:.bazelrc`. Because this
`//:.bazelrc.user` file is loaded at the end of `//:.bazelrc`, options
specified in the former will override those set in the latter.

Closes #190

GitOrigin-RevId: 139c6dded55f981cc45add69004e58d7dfc9e969
Change-Id: Id89084d29fd1e56a65183544573d24a642032357