git integrate: Don't merge unrelated repos
If copybara integrate label references a non related repo, copybara
will blindly create a merge (ignoring the files) to the other repo.

The consequence is that the repo size can increase size dramatically
as it is referring a commit of a whole different repo.

Also fixed some tests that were using this feature for make the test
simpler. We were using unrelated repos to show that the merge work.
Now we create a common parent for both repos.

INCLUDE_FILES logic is kept untouched since it doesn't create merges.

FIXED=228310119

Change-Id: I64b35e63eb29cce592c7a47fedf3a91bc6f55660