RE2 regex module
RE2 regexes are used extensively thru Starlark config (e.g.
core.replace), but cannot be used without being in the context
of a transformation or specific object.

This change adds the ability to use RE2 regexes in dynamic
transformations, git.mirror actions and in general any code
that uses Starlark.

The way it is exposed is a thin Starlark wrapper of the API.
This API is very stable as it is the same as the standard
java one.

Change-Id: Id48c3c030d61caa071854dc5cd95960db80fbc27