Exclude directories instead of individual files
This is a retry of a529b9df80

This change uses the ideas proposed in #191 (Thanks so much
aaron-skydio!!).

The difference with a529b9df80 is that that change didn't take into
account that after 'git add' but before exclude reverts is run, new
files (included files) are deleted from the workdir and are only
present in the index.

The approach in this change is to 1) look for the file tree in the
HEAD commit, 2) look at the current workdir. From 1) get the
exclude list and combine 1) and 2) data to get all included files,
including deletes. Then 3) we only allow directory exclusions that
don't overlap with included files and their parents.

Also move everything to relative paths since we can batch more files
per batch git add.

TreeSet for excludes (ordered) gave me a slightly faster performance
(23s vs 24s). But also makes the logs easier to understand.

Fix #191

PiperOrigin-RevId: 451396225
Change-Id: Iff4073c627f1be683318b33ad26e792c02c5f4ba