Allow StarlarkValue#str to access StarlarkSemantics
This allows us to alter the behavior of `str(some_starlark_value)` depending on flags. See child CL for an example.

This involves adding the StarlarkSemantics argument to a rather long list of call sites and implementors, but it helps in the end to clarify that stringification is really dependent on starlark semantics.

PiperOrigin-RevId: 461852476
Change-Id: Ie826e67987c440ed818546ce020fcdd80d252f64