Introduce new "mergeImport" mode
mergeImport mode merges origin files into the destination using diff3
and a common baseline.

The purpose of this mode is to allow users to perpetuate
destination-only changes. Typically, Copybara only allows edits
directly to the source of truth (origin). However, users often want to
maintain files (and portions of files) only in the destination. This
workflow mode is meant to eliminate the need for patch files for this
type of workflow by merging the origin files into the destination.

To adequately test, we need to add a DestinationReader implementation to
RecordsProcessCallDestination so we can call copyDestinationFiles().

Additionally, this CL contains a few other small changes:
- RemoteArchiveRevision gives correct value for asString()
- introduce new param to use as origin baseline for merge import mode

Future changes will clarify that we accept any binary for performing
the merging in merge_input and will create a new working directory
for the output that binary. We will use this to prevent the merged file
from overwriting the original checkout file.

BUG=213348659,241409625
PiperOrigin-RevId: 466414452
Change-Id: I30cb32898f7e1d040d28ed39ca648757324868b7